LPS-92129 adapt
